A-MoreOOP: Use More OOP
